pyproject.toml
[project]
name = "shiftml"
version = "0.0.0"
requires-python = ">=3.9"
authors = [
]
dependencies = [
    "ase",
    "platformdirs",
    "metatensor-learn==0.3.1",
    "metatensor-operations==0.3.1",
    "metatensor-torch==0.7.3",
    "vesin",
]

readme = "README.md"
license = {text = "BSD-3-Clause"}
description = "Predictions of chemical shieldings using machine learning"

[project.urls]
homepage = "https://github.com/lab-cosmo/ShiftML"
# documentation = "TODO"
repository = "https://github.com/lab-cosmo/ShiftML"



[build-system]
requires = [
    "setuptools >=68",
]
build-backend = "setuptools.build_meta"
